fix deployment to execute flask app
